client renamed to py-client
